moved all config files into a single config dir
