+++ components/madewithcode-intro.jsx
@@ -1,7 +1,7 @@
 var React = require('react');
 var ImageTag = require('../components/imagetag.jsx');
 
-var config = require('../lib/build/config');
+var config = require('../config/config');
 
 var MadeWithCodeIntro = React.createClass({
   propTypes: {
