+++ config/index.static.config.js
