+++ gulp/shared/build.tasks.js
@@ -1,2 +1,1 @@
-var MINIMAL_BUILD_TASKS = require('./minimal.build.tasks.js');
-module.exports = MINIMAL_BUILD_TASKS.concat(['sitemap']);
+module.exports = require('./minimal.build.tasks.js').concat(['sitemap']);
