+++ lib/blog-feed-loader.js
@@ -1,4 +1,4 @@
-var config = require('./build/config');
+var config = require('../config/config');
 var testFeeder = require('../test/browser/stub-blog-feed-loader');
 var googleFeeder = require('./google-blog-feed-loader');
