+++ pages/about.jsx
@@ -5,7 +5,7 @@ var IconLinks = require('../components/icon-links.jsx');
 var IconLink = require('../components/icon-link.jsx');
 var Illustration = require('../components/illustration.jsx');
 
-var config = require('../lib/build/config');
+var config = require('../config/config');
 
 var Intro = React.createClass({
   render: function () {
