+++ pages/opportunities.jsx
@@ -5,7 +5,7 @@ var HeroUnit = require('../components/hero-unit.jsx');
 var Illustration = require('../components/illustration.jsx');
 var IconLinks = require('../components/icon-links.jsx');
 var IconLink = require('../components/icon-link.jsx');
-var config = require('../lib/build/config');
+var config = require('../config/config');
 
 var SubSection = React.createClass({
   render: function() {
