+++ pages/private-eye.jsx
@@ -4,7 +4,7 @@ var Instructions = require('../components/madewithcode-instructions.jsx');
 var Router = require('react-router');
 var Link = Router.Link;
 
-var config = require('../lib/build/config');
+var config = require('../config/config');
 
 var Lightbeam = React.createClass({
   render: function() {
