+++ test/config.test.js
@@ -1,6 +1,6 @@
 var should = require('should');
 
-var config = require('../lib/build/config');
+var config = require('../config/config');
 
 describe('config', function() {
   it('should think it is in a test suite', function() {
